ghcid auto-reload for tests
This change extends multi-home units to work with tests as well. We
first make vira tests a proper spec (not executable), and then add that
to --enable-multi-repl so that the tests re-compile when the library
sources change.

This option is define in one place, ./cabal-repl. We could potentially
nixify this using https://github.com/srid/haskell-flake/pull/449

Also, the tests no longer build in Nix for vira, as they require
network. Instead the Github workflow runs them through cabal.