refactor: Move logging to Vira.Lib (#107)
The logging part doesn't require application domain, so it shouldn't be
in Vira.App.

Incidentally, this makes Vira.Supervisor independent of the rest of the
app (only dependent on Vira.Lib.*)

Thus, we could potentially make the supervisor its own package (in
`./packages`) but I will not do that while #83 is in flight.